// src/app/shared/model/attendance.ts
export interface attendance {
    uid: string
    login_date_time: string
    day_count: string
    logout_date_time: string
    work_from: string
}
export interface attendance_List {
    id: string
    uid: string
    login_date_time: string
    day_count: string
    logout_date_time: string
    work_from: string
}